+++ NEWS.md
@@ -1,4 +1,4 @@
 # nyschooldata 0.0.1
 
 * initial commits
-* assessment database, 2014 and 2015
+* assessment database, 2013-2015, using helper in Hmisc